requirements.txt
# Core ML and Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# MLOps and Tracking
mlflow>=2.8.0
evidently>=0.4.0

# API and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.25.0

# Configuration and Utils
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Download
kaggle>=1.5.16

# Cloud Storage (optional)
google-cloud-storage>=2.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
types-PyYAML>=6.0.12

# Pre-commit hooks (optional)
pre-commit>=3.5.0

